#554: move metadata handler to src/metadata
