+++ src/metadata/metadata_handler.cc
