+++ src/metadata/metadata_handler.h
