+++ src/upnp_xml.cc
@@ -32,7 +32,7 @@
 #include "upnp_xml.h"
 #include "common.h"
 #include "config_manager.h"
-#include "metadata_handler.h"
+#include "metadata/metadata_handler.h"
 #include "server.h"
 
 using namespace zmm;
