+++ src/util/tools.cc
@@ -59,7 +59,7 @@
 #include "config_manager.h"
 #include "contrib/md5.h"
 #include "iohandler/file_io_handler.h"
-#include "metadata_handler.h"
+#include "metadata/metadata_handler.h"
 #include "tools.h"
 
 #define WHITE_SPACE " \t\r\n"
