+++ src/web/add_object.cc
@@ -32,7 +32,7 @@
 #include "cds_objects.h"
 #include "common.h"
 #include "content_manager.h"
-#include "metadata_handler.h"
+#include "metadata/metadata_handler.h"
 #include "pages.h"
 #include "server.h"
 #include "util/tools.h"
